Add nice-citation to `gnus-treatment-function-alist`
Thanks to Eric Abrahamsen, this is a very clean way to apply
nice-citation in article-mode only once. It can be disabled with
custom variable `nice-citation-treat-citations`, and it is not applied
when viewing an article with C-u g and C-u C-u g, as should be!

To make it work properly, had to ensure nice-citation starts
from (point-min), which apparently was not always the case.

The abusing of `fontification-functions` is still present for
message-mode. I'll see if I can find a cleaner way there too...